CP 3-parameter hyperbolic model

Fits critical power with the 3-parameter hyperbolic model (Morton, 1996). For more details see ?critical_power.

method_3_hyp(
  .data,
  power_output_column,
  time_to_exhaustion_column,
  plot = TRUE,
  power_in_y_axis = FALSE
)

Arguments

.data
A data frame containing a power output column and their respective time-to-exhaustion.

power_output_column
The name of the power output column. This value has to be in watts. Default to "PO".

time_to_exhaustion_column
The name of the time-to-exhaustion column. This value has to be in seconds. Default to "TTE".

plot
A boolean indicating whether to produce a plot from the critical power estimation. Default to TRUE.

power_in_y_axis
A boolean to indicate whether to plot the Power Output in the y-axis. Default to FALSE.

Value

a tibble containing the following columns:

method

The critical power method for that estimation.

data

The data you provided.

model

A nls or lm object. The model used in the critical power fitting.

CP

Critical power, in watts.

CP SEE

Standard error of the estimation of critical power, in watts.

W'

Anaerobic work capacity, in joules.

W' SEE

Standard error of the estimation of anaerobic work capacity, in joules.

Pmax

Maximal instantaneous power, in watts (only available in CP 3-hyp).

Pmax SEE

Standard error of the estimation of maximal instantaneous power, in watts (only available in CP 3-hyp).

R2

R-squared.

RMSE

Root mean square error, the units vary depending on the method, as they represent the units of the response (i.e., y-axis): CP3-hyp in seconds, CP2-hyp in seconds, CPlinear in joules, and CP1/time in watts.

plot

The critical power plot.

Examples

results_3_hyp <- method_3_hyp(.data = demo_data, "PO", "TTE")

results_3_hyp

#> # A tibble: 1 x 11
#>   data   model    CP `CP SEE`   `W'` `W' SEE`  Pmax `Pmax SEE`    R2  RMSE plot 
#>   <list> <lis> <dbl>    <dbl>  <dbl>    <dbl> <dbl>      <dbl> <dbl> <dbl> <lis>
#> 1 <tibb… <nls>  260.      3.1 27410.     4794 1004.       835. 0.998  37.1 <gg>