CP models

These functions fit the following critical power/speed models:

3-parameter hyperbolic model (Morton, 1996).

2-parameter hyperbolic model (Hill, 1993).

linear model (Moritani, 1981).

1/time linear model (Whipp et al. 1982).
For more details see ?critical_power or ?critical_speed.

method_cp(
  .data,
  power_output_column,
  distance_column,
  time_to_exhaustion_column = "TTE",
  method = c("3-hyp", "2-hyp", "linear", "1/time"),
  mode = c("power", "speed"),
  plot = TRUE,
  reverse_y_axis = FALSE
)

Arguments

.data
A data frame containing a power output/distance column and their respective time-to-exhaustion.

power_output_column
The name of the power output column. This value has to be in watts. Used in the 'power' mode only.

distance_column
The name of the distance column. This value has to be in meters. Used in the 'speed' mode only.

time_to_exhaustion_column
The name of the time-to-exhaustion column. This value has to be in seconds. Default to "TTE".

method
The method for estimating critical speed. It can be one or multiple methods. Default to c("3-hyp", "2-hyp", "linear", "1/time").

mode
Specify 'power' for critical power, and 'speed' for critical speed.

plot
A boolean indicating whether to produce a plot from the critical power/speed estimation. Default to TRUE.

reverse_y_axis
A boolean to indicate whether to plot the Power Output (W) or Speed (m/s) in the y-axis. It is ignored for the linear methods. Default to FALSE.

Value

a tibble containing the following columns:

method

The critical power method for that estimation.

data

The data you provided.

model

A nls or lm object. The model used in the critical power fitting.

CP or CS

Critical power, in watts, or critical speed, in meters per second.

CP SEE or CS SEE

Standard error of the estimation of critical power, in watts, or critical speed, in meters per second.

W' or D'

Anaerobic work capacity, in joules, or anaerobic distance capacity, in meters.

W' or D' SEE

Standard error of the estimation of anaerobic work capacity, in joules, or anaerobic distance capacity, in meters.

Pmax or Smax

Maximal instantaneous power/speed, in watts/meters per second (only available in CP/CS 3-hyp).

Pmax SEE or Smax SEE

Standard error of the estimation of maximal instantaneous power/speed, in watts/meters per second (only available in CP/CS 3-hyp).

R2

R-squared.

RMSE

Root mean square error, the units vary depending on the method, as they represent the units of the response (i.e., y-axis): CP3-hyp in seconds, CP2-hyp in seconds, CPlinear in joules, and CP1/time in watts.

plot

The critical power plot.