Plot CP model

Plot CP model

plot_cp(
  .data,
  power_output_column,
  time_to_exhaustion_column,
  method = c("3-hyp", "2-hyp", "linear", "1/time"),
  model
)

Arguments

.data
A data frame containing a power output column and their respective time-to-exhaustion.

power_output_column
The name of the power output column. This value has to be in watts. Default to "PO".

time_to_exhaustion_column
The name of the time-to-exhaustion column. This value has to be in seconds. Default to "TTE".

method
The method for estimating critical power. It can be one of c("3-hyp", "2-hyp", "linear", "1/time").

model
The nls or lm object derived in one of the method_ functions.

Value

A ggplot2 object